Rename classes Application, Semaphore, Version => Model::Application and under
